// src/utils/types.tsx
import PropTypes from 'prop-types';
import { BasePlacement } from '@popperjs/core'

export type Variant =
  | 'primary'
  | 'secondary'
  | 'success'
  | 'danger'
  | 'warning'
  | 'info'
  | 'dark'
  | 'light'
  | string;
export type ButtonVariant =
  | Variant
  | 'link'
  | 'outline-primary'
  | 'outline-secondary'
  | 'outline-success'
  | 'outline-danger'
  | 'outline-warning'
  | 'outline-info'
  | 'outline-dark'
  | 'outline-light';
export type Color =
  | 'primary'
  | 'secondary'
  | 'success'
  | 'danger'
  | 'warning'
  | 'info'
  | 'dark'
  | 'light'
  | 'white'
  | 'muted';

export type Placement = import('@restart/ui/usePopper').Placement;
export type TooltipPlacement = BasePlacement | "top-start" | "bottom-start"
export type CardVariant = 'card-action' | 'card-action-quantity-toggle'

export type AlignDirection = 'start' | 'end';

export type ResponsiveAlignProp =
  | { sm: AlignDirection }
  | { md: AlignDirection }
  | { lg: AlignDirection }
  | { xl: AlignDirection }
  | { xxl: AlignDirection };

export type AlignType = AlignDirection | ResponsiveAlignProp;
export const alignDirection = PropTypes.oneOf<AlignDirection>(['start', 'end']);

export const alignPropType = PropTypes.oneOfType([
  alignDirection,
  PropTypes.shape({ sm: alignDirection }),
  PropTypes.shape({ md: alignDirection }),
  PropTypes.shape({ lg: alignDirection }),
  PropTypes.shape({ xl: alignDirection }),
  PropTypes.shape({ xxl: alignDirection }),
]);

export type RootCloseEvent = 'click' | 'mousedown';

export type GapValue = 0 | 1 | 2 | 3 | 4 | 5;

export type Size = 'sm' | 'md' | 'lg' | 'xl' | 'xxl'